Fix unvalid IDs making created items unsellable
Right now the items created using the profile editor are unsellable because the vendors will notice the that the ids are invalid because they contain non-hexadecimal characters. 
I propose fixing it by stop using a "I" at the start of the IDs and make the "sign" that's appended at the end of the ids contain only hexadecimal characters.